// frontend/src/lib/api/generated/models/ListingAttendeesRankingItem.ts
/**
 * Roominar
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class ListingAttendeesRankingItem {
    'id': number;
    'fullName': string;
    'email': string;
    'avatarUrl'?: string | null;
    'totalScore': number;
    'purchaseNumber'?: number;
    'checkinNumber'?: number;
    'surveyNumber'?: number;
    'prevScore'?: number | null;
    'rankChange': ListingAttendeesRankingItemRankChangeEnum;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "number",
            "format": ""
        },
        {
            "name": "fullName",
            "baseName": "full_name",
            "type": "string",
            "format": ""
        },
        {
            "name": "email",
            "baseName": "email",
            "type": "string",
            "format": ""
        },
        {
            "name": "avatarUrl",
            "baseName": "avatar_url",
            "type": "string",
            "format": ""
        },
        {
            "name": "totalScore",
            "baseName": "total_score",
            "type": "number",
            "format": ""
        },
        {
            "name": "purchaseNumber",
            "baseName": "purchase_number",
            "type": "number",
            "format": ""
        },
        {
            "name": "checkinNumber",
            "baseName": "checkin_number",
            "type": "number",
            "format": ""
        },
        {
            "name": "surveyNumber",
            "baseName": "survey_number",
            "type": "number",
            "format": ""
        },
        {
            "name": "prevScore",
            "baseName": "prev_score",
            "type": "number",
            "format": ""
        },
        {
            "name": "rankChange",
            "baseName": "rank_change",
            "type": "ListingAttendeesRankingItemRankChangeEnum",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return ListingAttendeesRankingItem.attributeTypeMap;
    }

    public constructor() {
    }
}


export enum ListingAttendeesRankingItemRankChangeEnum {
    Up = 'up',
    Down = 'down',
    Same = 'same',
    New = 'new'
}